refactor: move weatherMonitor to own module
