+++ src/index.js
@@ -3,10 +3,10 @@
 
 const hid = require("node-hid");
 const os = require("os-utils");
-const request = require("request");
 
 const { wait } = require("./util");
 const perfStream = require("./perfMonitor").stream;
+const weatherStream = require("./weatherMonitor").stream;
 const stockStream = require("./stockMonitor").stream;
 
 
@@ -25,78 +25,6 @@ let currentScreenIndex = 0;
 let keyboard = null;
 let screenBuffer = null;
 let screenLastUpdate = null;
-
-async function startWeatherMonitor() {
-    // Regex's for reading out the weather info from the yahoo page
-    const tempRegex = /"temperature":({[^}]+})/;
-    const condRegex = /"conditionDescription":"([^"]+)"/;
-    const rainRegex = /"precipitationProbability":([^,]+),/;
-
-    function getWeather() {
-        return new Promise((resolve) => {
-            request(`https://www.yahoo.com/news/weather/united-states/st-augustine/st-augustine-12771497`, (err, res, body) => {
-                const weather = {};
-                const temp = tempRegex.exec(body);
-                if (temp && temp.length > 1) {
-                    weather.temp = JSON.parse(temp[1]);
-                }
-
-                const cond = condRegex.exec(body);
-                if (cond && cond.length > 1) {
-                    weather.desc = cond[1];
-                }
-
-                const rain = rainRegex.exec(body);
-                if (rain && rain.length > 1) {
-                    weather.rain = rain[1];
-                }
-                resolve(weather);
-            });
-        });
-    }
-
-    // Used for scrolling long weather descriptions
-    let lastWeather = null;
-    let lastWeatherDescIndex = 0;
-
-    // Just keep updating the data forever
-    while (true) {
-        // Get the current weather for Seattle
-        const weather = await getWeather();
-        if (weather && weather.temp && weather.desc && weather.rain) {
-            let description = weather.desc;
-
-            // If we are trying to show the same weather description more than once, and it is
-            // longer than 9
-            // Which is all that will fit in our space, lets scroll it.
-            if (lastWeather && weather.desc == lastWeather.desc && weather.desc.length > 9) {
-                // Move the string one character over
-                lastWeatherDescIndex++;
-                description = description.slice(lastWeatherDescIndex, lastWeatherDescIndex + 9);
-                if (lastWeatherDescIndex > weather.desc.length - 9) {
-                    // Restart back at the beginning
-                    lastWeatherDescIndex = -1; // minus one since we increment before we show
-                }
-            } else {
-                lastWeatherDescIndex = 0;
-            }
-            lastWeather = weather;
-
-            // Create the new screen
-            const screen =
-                `desc: ${ description }${ " ".repeat(Math.max(0, 9 - ("" + description).length)) } |  ${ title(0, 2) } ` +
-                `temp: ${ weather.temp.now }${ " ".repeat(Math.max(0, 9 - ("" + weather.temp.now).length)) } |  ${ title(1, 2) } ` +
-                `high: ${ weather.temp.high }${ " ".repeat(Math.max(0, 9 - ("" + weather.temp.high).length)) } |  ${ title(2, 2) } ` +
-                `rain: ${ weather.rain }%${ " ".repeat(Math.max(0, 8 - ("" + weather.rain).length)) } |  ${ title(3, 2) } `;
-
-            // Set this to be the latest weather info
-            screens[SCREEN_WEATHER] = screen;
-        }
-
-        // Pause a bit before requesting more info
-        await wait(KEYBOARD_UPDATE_TIME);
-    }
-}
 
 async function sendToKeyboard(screen) {
     // If we are already buffering a screen to the keyboard just quit early.
